🍎 ViewModel 구조 변경
- 각 Screen 마다 별도의 ViewModel 을 사용하기 위해 HomeScreen ViewModel 생성
- MainViewModel 의 경우 Network 관리에 초점을 맞춤
