+++ app/src/main/java/com/example/moviesearchapp/view/MainActivity.kt
@@ -3,31 +3,30 @@ package com.example.moviesearchapp.view
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.lifecycleScope
 import com.example.moviesearchapp.ui.theme.ComposeMovieSearchAppTheme
 import com.example.moviesearchapp.view.navigation.MovieNavigation
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.CoroutineScope
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+    private val mainViewModel: MainViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             ComposeMovieSearchAppTheme {
-                MovieApp(coroutineScope = lifecycleScope)
+                MovieApp(mainViewModel)
             }
         }
     }
 }
 
 @Composable
-fun MovieApp(coroutineScope: CoroutineScope) {
-    val viewModel: MainViewModel = hiltViewModel()
-    MovieNavigation(mainViewModel = viewModel, coroutineScope = coroutineScope)
+fun MovieApp(mainViewModel: MainViewModel) {
+    MovieNavigation(mainViewModel = mainViewModel)
 }
 
 @Preview(showBackground = true)
