+++ app/src/main/java/com/example/moviesearchapp/view/MainViewModel.kt
@@ -1,25 +1,9 @@
 package com.example.moviesearchapp.view
 
 import androidx.lifecycle.ViewModel
-import com.example.domain.usecase.RequestMovieListUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor(
-    private val requestMovieListUseCase: RequestMovieListUseCase
-) : ViewModel() {
-
-    private val _searchQuery = MutableStateFlow("")
-    val searchQuery = _searchQuery.asStateFlow()
-
-    fun setSearchQuery(query: String) {
-        _searchQuery.value = query
-    }
-
-    fun onClearQuery() {
-        _searchQuery.value = ""
-    }
+class MainViewModel @Inject constructor() : ViewModel() {
 }
